Rename launch_command to open_program & move it to desktop_file.rs
