+++ src/config/mod.rs
@@ -44,11 +44,9 @@ use crate::modules::workspaces::WorkspacesModule;
 
 use crate::modules::{AnyModuleFactory, ModuleFactory, ModuleInfo};
 use cfg_if::cfg_if;
-use color_eyre::{Help, Report, Result};
+use color_eyre::Result;
 use serde::Deserialize;
 use std::collections::HashMap;
-use std::process::{Command, Stdio};
-use tracing::error;
 
 #[cfg(feature = "schema")]
 use schemars::JsonSchema;
@@ -440,22 +438,4 @@ pub const fn default_true() -> bool {
 
 pub fn default_launch_command() -> String {
     String::from("gtk-launch {app_name}")
-}
-
-pub fn launch_command(file_name: &str, str: &str) {
-    let expanded = str.replace("{app_name}", file_name);
-    let launch_command_parts: Vec<&str> = expanded.split_whitespace().collect();
-    if let Err(err) = Command::new(&launch_command_parts[0])
-        .args(&launch_command_parts[1..])
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-    {
-        error!(
-            "{:?}",
-            Report::new(err)
-                .wrap_err("Failed to run launch command.")
-                .suggestion("Perhaps the applications file is invalid?")
-        );
-    }
 }
